Rely on `autopub check` to halt Travis
Instead of having `autopub check` output "true" (just for Travis) when a
release file is found, setting a HAS_RELEASE env var with that value,
and then running the `deploy` stage only if that var is true, this new
flow depends on `autopub check` sending `travis_terminate 0` instead.

This also dispenses with the `after_success` stage, instead putting all
pre-deployment steps in the `before_deploy` stage.